Add editting and insert backend (MongoDB)
Add editting and insert backend (MongoDB)
Try to add "onclick" event with ID to the "Save change" button but a bug : if you click on multiple pen to change value but close the modal, when you relly want to edit and save, it save all the id where you have click on edditing before (due to the addListener).
Correction : set in script the ID when modal open and get in script the ID when click to save change(s).